+++ talk2kobold.py
@@ -24,7 +24,7 @@ default_prompt = {
     "use_world_info": False,  #?
     "quiet": True,
     "max_context_length": 2048,
-    "max_length": 10,
+    "max_length": 16,
     "singleline": False,  #?
     "n": 1,
     "temperature": 0.7,
@@ -57,6 +57,7 @@ def eval_template(template, context):
 class Conversation:
 
     endpoint = "http://127.0.0.1:5001"
+    cutoff_digits = 8
 
     def __init__(self, user, bot=None):
         self.user = user
@@ -71,75 +72,86 @@ class Conversation:
         self.bot = bot
         memory = "\n".join((bot["persona"], bot["scenario"], bot["example"]))
         self.prompt_data["memory"] = self.parse_vars(memory)
-        self.prompt_file = f"aiclient.log"
-        self.log = f"aiclient_full.log"
-        self.prompt = ""
-        self.load_prompt()
+        self.memory_tokens = self.count_tokens(memory)
+        self.log = f"aiclient.log"
+        self.load_history()
 
     def parse_vars(self, text):
         context = dict(user=self.user, char=self.botname)
         return eval_template(text, context)
 
-    def load_prompt(self):
-        with open(self.prompt_file, 'a+') as file:
+    def load_history(self):
+        with open(self.log, "a+") as file:
             file.seek(0)
-            self.prompt = file.read()
-        if self.prompt == "":
-            self.to_prompt( self.parse_vars(self.bot["first_mes"]) )
+            text = file.read()
+            field = text[-self.cutoff_digits:]
+            if field.isdecimal():
+                self.cutoff = int(field)
+                self.prompt = text[:-self.cutoff_digits-1]
+            else:
+                tolog("\n#### Warning: no cutoff in log! ####\n\n")
+                self.cutoff = 0
+                self.prompt = text
+                file.write(f"\n{self.cutoff:0{self.cutoff_digits}}")
+        self.to_prompt("")  # shift context, if log was extended manually
         print(self.prompt,"\n", sep="")
 
-    def to_history(self, text):
-        if text != "":
-            with open(self.log, "a") as f:
-                f.write(text)
-
-    def check_saved(self):
-        with open(self.prompt_file, "r") as f:
-            saved = f.read()
-        if self.prompt != saved:
-            tolog(f'save corruption:\nsaved=[{saved}]\norig=[{self.prompt}]\n')
-
-    def load_history(self):
-        with open(self.log) as f:
-            text = f.read()
-        self.cutoff = int(text[-8:])
-        self.history = text[:-8]
-
-    def save_message(self, msg):
+    def to_history(self, msg):
         with open(self.log, "r+") as f:
-            f.seek(-8, os.SEEK_END)
+            f.seek(0, os.SEEK_END)
+            end = f.tell()
+            if end > self.cutoff_digits:
+                f.seek(end-self.cutoff_digits-1)
             f.write(msg)
-            f.write(f"{self.cutoff:8}")
+            f.write(f"\n{self.cutoff:0{self.cutoff_digits}}")
+
+    # up to 300 tokens shifts were observed, we can assume no small limit there
+    def shift_context(self, shift):
+        pos = self.cutoff+shift*5+100
+#        for end in ('\n\n', '.\n', '"\n', '\n', ' '):
+#            pos2 = self.prompt.find(end, pos, pos+400)
+#            if pos2 >= 0:
+#                pos = pos2+len(end)
+#                break
+        tolog(f'{shift=} cutting={pos-self.cutoff}\n\ncut=[{self.prompt[self.cutoff:pos]}]\n\nnew_start=[{self.prompt[pos:pos+60]}]...\n')
+        self.cutoff = pos
+
+    """
+    def to_prompt(self, message):
+        self.prompt += message
+        now = len(self.prompt) - self.cutoff
+        #!!! if 30//10 will work, try 29//10. 30//10 with 4k context may overflow.
+        max_context = self.prompt_data['max_context_length']
+        max = (max_context-self.memory_tokens)*30//10
+        add = self.count_tokens(self.prompt)
+        # - len(self.prompt_data["memory"])
+        tolog(f'{now} < {max} = {len(self.prompt_data["memory"])}\n')
+        if now > max:
+            tolog(f'message {len(message)}=[{message}]\n')
+            self.shift_context(now-max)
+        elif message == "":
+            return
+        self.to_history(message)
+    """
 
     def to_prompt(self, message):
-        self.to_history(message)
         self.prompt += message
-        now = len(self.prompt)
-        max = (self.prompt_data['max_context_length']*3
-            - len(self.prompt_data["memory"]))
-        tolog(f'{now} < {max} = {self.prompt_data["max_context_length"]*28//10}-{len(self.prompt_data["memory"])}\n')
-        if now > max:
-            pos = now-max
-            for end in ('\n\n', '.\n', '"\n', '\n'):
-                pos2 = self.prompt.find(end, pos)
-                if pos2 < 120 or end == '\n':
-                    pos = pos2
-                    break
-            tolog(f'cut=[{self.prompt[:pos+len(end)]}]\nnew_start=[{self.prompt[pos+len(end):pos+200]}]\n')
-            self.prompt = self.prompt[pos+len(end):]
-            with open(self.prompt_file, "w") as f:
-                f.write(self.prompt)
-            self.check_saved()
-        elif message != "":
-            with open(self.prompt_file, "a") as f:
-                f.write(message)
-            self.check_saved()
+        max = self.prompt_data['max_context_length'] - self.memory_tokens
+        now = self.count_tokens(self.prompt[self.cutoff:])
+        extra = now-max+10+self.prompt_data['max_length']
+        tolog(f'tokens: {extra=}, {now} < {max}, memory={self.memory_tokens}\n')
+        if extra > 0:
+            tolog(f'message {len(message)}=[{message}]\n')
+            self.shift_context(extra)
+        elif message == "":
+            return
+        self.to_history(message)
 
     def get_json_prompt(self):
         context = dict(user=self.user, char=self.botname)
         stop = [eval_template(s, context) for s in self.stop_sequence]
         self.prompt_data["stop_sequence"] = stop
-        self.prompt_data["prompt"] = self.prompt
+        self.prompt_data["prompt"] = self.prompt[self.cutoff:]
         return self.prompt_data
 
     def send_to_ai(self):
@@ -149,14 +161,15 @@ class Conversation:
             raise IOError  #!!!todo add text
         results = response_data.json()['results']
         response = results[0]['text']
-        return response
+        status = self.status()
+        return response, status['stop_reason']
 
     def collect_response(self, message):
         self.to_prompt(message)
         generated = 0
         done = False
         while True:
-            response = self.send_to_ai()
+            response, stop_reason = self.send_to_ai()
             generated += len(response)
             for suffix in self.prompt_data['stop_sequence']:
                 if response.endswith(suffix):
@@ -165,11 +178,19 @@ class Conversation:
                         response += "\n"
                     done = True
             if not done:
-                if ( len(response) < 1*self.prompt_data["max_length"] or
+                if stop_reason == 1:
+                    response = response.rstrip()+"\n"
+                    done = True
+                elif ( len(response) < 1*self.prompt_data["max_length"] or
                         response.rstrip()[-2:] in ('."', '!"', '?"', ',"') ):
                     response = response.rstrip()+"\n"
                     done = True
                 elif generated > 240:
+                    pos = response.rfind(".")
+                    if pos >= 0:
+                        response = response[:pos+1]
+                        done = True
+                elif generated > 160:
                     pos = response.rfind("\n")
                     letter = response[pos-1:pos]
                     if pos >= 0 and letter in '.!?"':
@@ -192,6 +213,27 @@ class Conversation:
         except IOError:
             print("Error: can not send message.")
 
+    def abort(self):
+        requests.post(f"{self.endpoint}/api/extra/abort")
+
+    def status(self):
+        """ Get status of KoboldCpp
+        Result: last_process, last_eval, last_token_count, total_gens, queue, idle,
+        stop_reason (INVALID=-1, OUT_OF_TOKENS=0, EOS_TOKEN=1, CUSTOM_STOPPER=2)
+        """
+        response = requests.get(f"{self.endpoint}/api/extra/perf")
+        if response.status_code != 200:
+            raise IOError  #!!!todo add text
+        return response.json()
+
+    def count_tokens(self, text):
+        response = requests.post(f"{self.endpoint}/api/extra/tokencount",
+            json={"prompt": text})
+        if response.status_code != 200:
+            tolog(f"count_tokens: err={response.status_code}\n")
+            raise IOError  #!!!todo add text
+        return response.json()['value']
+
 
 def format_bot(bot):
     for k in bot:
@@ -201,6 +243,7 @@ def format_bot(bot):
 def talk(bot):
     format_bot(bot)
     chat = Conversation("You", bot)
+    chat.abort()
     while True:
         user_message = input(f"{chat.user}> ")
         print()