+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## 2023.12.16
 
+- Add extended kobold api.
+- Prepare token counting.
+- Get stop_reason from kobold engine.
+- Single history variable and history file with cutoff.
+
+## 2023.12.16
+
 - Development version with a lot of garbage.
 - Prompt file with stored cutoff.
 - Embedded assistant bot.